+++ FilesConverter/FilesConverter/StocksConverter/VentaStocksConverter.cs
@@ -27,7 +27,7 @@ namespace FilesConverter.StocksConverter
                 if (Helper.IsRowEmpty(row)) continue;
                 for (int j = 2; j < columns.Count; j++)
                 {
-                    int i;
+                    decimal i;
                     var storedSalesRow = new StocksResultItem
                     {
                         Customer = Customer,
@@ -35,7 +35,7 @@ namespace FilesConverter.StocksConverter
                         AdressSklada = "Вента" + " " + columns[j].Caption,
                         Date = Date.Date,
                         ItemName = row["Наименование Препарата"].ToString(),
-                        Upakovki = int.TryParse(row[columns[j].Caption].ToString(), out i) ? i : (int?)null
+                        Upakovki = decimal.TryParse(row[columns[j].Caption].ToString(), out i) ? i : (decimal?)null
 
                     };
                     commonResultLines.Add(storedSalesRow);
